TRACKING: Cleanup / simplify components

+++ src/lib/kotlin/slatekit-tracking/src/main/kotlin/slatekit/tracking/Expiry.kt
@@ -4,7 +4,7 @@ import slatekit.common.DateTime
 
 data class Expiry(val seconds:Long,
                   val started:DateTime = DateTime.now(),
-                  val expires:DateTime = started.plusSeconds(seconds.toLong())) {
+                  val expires:DateTime = started.plusSeconds(seconds)) {
 
     fun isExpired(): Boolean {
         return if (seconds <= 0) false else expires <= DateTime.now()
